requirements.txt
Pillow==2.2.1
sympy==0.7.3
